Refactor, make public the connection slot
There are some advanced features that require access to the underlying connection. The connection slot type will be exported by the sender interface.
