+++ conn/slot.go
